Use a combobox to select the icon
The old selection method (by name) was broken in Cinnamon 4.0 with
the introduction of the new xapp iconchooser. It only allows to select
existing icon names or paths.

This breaks backwards compatibility with 3.8 because of the introducion
of advanced dependency settings. The code is still the same though.

The extension has been split into two version folders. Cinn >= 3.8 uses
the extension inside the 3.8 folder and smaller versions use the
files located a the root.